Refactor : delete unused codes
