+++ src/main/java/com/palmseung/common/config/AppConfig.java
@@ -1,4 +1,4 @@
-package com.palmseung.config;
+package com.palmseung.common.config;
 
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
