+++ src/main/java/com/palmseung/members/jwt/JwtTokenProvider.java
@@ -1,4 +1,4 @@
-package com.palmseung.members.support;
+package com.palmseung.members.jwt;
 
 
 import io.jsonwebtoken.Claims;
