+++ src/main/java/com/palmseung/members/jwt/LoginUser.java
@@ -1,4 +1,4 @@
-package com.palmseung.members.support;
+package com.palmseung.members.jwt;
 
 import org.springframework.security.core.annotation.AuthenticationPrincipal;
