+++ src/main/java/com/palmseung/members/jwt/TokenProperties.java
@@ -1,4 +1,4 @@
-package com.palmseung.members.support;
+package com.palmseung.members.jwt;
 
 import lombok.Getter;
 import lombok.NoArgsConstructor;
