+++ src/main/java/com/palmseung/members/jwt/UserMember.java
@@ -1,4 +1,4 @@
-package com.palmseung.members.support;
+package com.palmseung.members.jwt;
 
 import com.palmseung.members.domain.Member;
 import lombok.Getter;
