+++ src/test/java/com/palmseung/keywords/docs/KeywordDocumentationTest.java
@@ -5,8 +5,8 @@ import com.palmseung.keywords.dto.MyKeywordRequestView;
 import com.palmseung.keywords.service.KeywordService;
 import com.palmseung.members.domain.Member;
 import com.palmseung.members.service.MemberService;
-import com.palmseung.members.support.JwtTokenProvider;
-import com.palmseung.members.support.UserMember;
+import com.palmseung.members.jwt.JwtTokenProvider;
+import com.palmseung.members.jwt.UserMember;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
