+++ src/test/java/com/palmseung/members/docs/MemberDocumentationTest.java
@@ -6,8 +6,8 @@ import com.palmseung.members.dto.CreateMemberRequestView;
 import com.palmseung.members.dto.LoginRequestView;
 import com.palmseung.members.dto.UpdateMemberRequestView;
 import com.palmseung.members.service.MemberService;
-import com.palmseung.members.support.JwtTokenProvider;
-import com.palmseung.members.support.UserMember;
+import com.palmseung.members.jwt.JwtTokenProvider;
+import com.palmseung.members.jwt.UserMember;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
