+++ src/test/java/com/palmseung/members/jwt/JwtTokenProviderTest.java
@@ -1,4 +1,4 @@
-package com.palmseung.members.support;
+package com.palmseung.members.jwt;
 
 import com.palmseung.members.domain.Member;
 import com.palmseung.members.domain.MemberRepository;
